refactor(webhook): restructure provider module, trim public surface
- Move WebhookRequest/Payload, WebhookResponse, WebhookContext, and the
  WebhookProvider trait into a new `provider/` module; `client/` now holds
  only the WebhookService wrapper and ServiceHealth. Crate-root re-exports
  are unchanged, so consumers are unaffected.
- Split WebhookContext into provider/context.rs.
- Remove unused WebhookRequest::into_payload (to_payload covers the use).
- Demote sign_payload and effective_user_agent to pub(crate).
- Fix crate-level TRACING_TARGET to "nvisy_webhook" for consistency.
