+++ crates/nvisy-webhook/src/client/mod.rs
@@ -1,25 +1,7 @@
-//! Webhook client types (requests, responses, health, and service wrapper).
+//! Webhook service wrapper and health types.
 
 mod health;
-mod request;
-mod response;
 mod service;
 
 pub use health::ServiceHealth;
-pub use request::{WebhookContext, WebhookPayload, WebhookRequest};
-pub use response::WebhookResponse;
 pub use service::WebhookService;
-
-use crate::Result;
-
-/// Core trait for webhook delivery operations.
-///
-/// Implement this trait to create custom webhook delivery providers.
-#[async_trait::async_trait]
-pub trait WebhookProvider: Send + Sync {
-    /// Delivers a webhook payload to the specified endpoint.
-    async fn deliver(&self, request: &WebhookRequest) -> Result<WebhookResponse>;
-
-    /// Performs a health check on the webhook provider.
-    async fn health_check(&self) -> Result<ServiceHealth>;
-}
